google-sheets-sync/requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
